criando testes para os serviços

+++ src/books-api/Books.Domain/Services/UserService.cs
@@ -4,6 +4,7 @@ using Books.Domain.Entities;
 using Books.Domain.Interfaces;
 using Books.Domain.Interfaces.Repositores;
 using Books.Domain.Interfaces.Services;
+using Books.Domain.Shared.Enums;
 using Books.Domain.Shared.Nofication;
 using Books.Domain.Shared.Resources;
 using MediatR;
@@ -13,10 +14,16 @@ namespace Books.Domain.Services
     public class UserService : Service, IUserService
     {
         private readonly IUserRepository _userRepository;
+        private readonly IRequestScope _requestScope;
 
-        public UserService(IUnitOfWork uow, IMediator bus, INotificationHandler<DomainNotification> notifications, IUserRepository userRepository) : base(uow, bus, notifications)
+        public UserService(IUnitOfWork uow,
+            IMediator bus,
+            INotificationHandler<DomainNotification> notifications,
+            IUserRepository userRepository,
+            IRequestScope requestScope) : base(uow, bus, notifications)
         {
             _userRepository = userRepository;
+            _requestScope = requestScope;
         }
 
         public void Add(UserDto dto)
@@ -27,6 +34,14 @@ namespace Books.Domain.Services
                 return;
             }
 
+            var currentUser = _requestScope.GetUser();
+
+            if(currentUser.Profile == ProfileType.Standard)
+            {
+                NotifyError(DomainError.StandardProfileUserDoesNotHavePermission);
+                return;
+            }
+
             var user = new User(dto.Name, dto.Password, dto.Email, dto.Profile.Value);
             _userRepository.Add(user);
             Commit();
@@ -34,12 +49,28 @@ namespace Books.Domain.Services
 
         public void Delete(Guid id)
         {
+            var currentUser = _requestScope.GetUser();
+
+            if (currentUser.Profile == ProfileType.Standard)
+            {
+                NotifyError(DomainError.StandardProfileUserDoesNotHavePermission);
+                return;
+            }
+
             _userRepository.Remove(id);
             Commit();
         }
 
         public void Disable(Guid id)
         {
+            var currentUser = _requestScope.GetUser();
+
+            if (currentUser.Profile == ProfileType.Standard)
+            {
+                NotifyError(DomainError.StandardProfileUserDoesNotHavePermission);
+                return;
+            }
+
             var user = _userRepository.GetById(id);
 
             if(user == null)
@@ -55,6 +86,14 @@ namespace Books.Domain.Services
 
         public void Enable(Guid id)
         {
+            var currentUser = _requestScope.GetUser();
+
+            if (currentUser.Profile == ProfileType.Standard)
+            {
+                NotifyError(DomainError.StandardProfileUserDoesNotHavePermission);
+                return;
+            }
+
             var user = _userRepository.GetById(id);
 
             if (user == null)
@@ -83,6 +122,14 @@ namespace Books.Domain.Services
 
         public void Update(Guid id, UserDto dto)
         {
+            var currentUser = _requestScope.GetUser();
+
+            if(currentUser.Profile == ProfileType.Standard && currentUser.Id != id)
+            {
+                NotifyError(DomainError.StandardProfileUserDoesNotHavePermission);
+                return;
+            }
+
             if (!dto.IsValid())
             {
                 NotifyValidationError(dto);
